requirements.txt
django==3.2.3
djangorestframework==3.12.4
django-oauth-toolkit==1.5.0
celery==5.1.0
psycopg2==2.8.6
clickhouse-driver==0.2.0
django-environ==0.4.5
